Move uuid.ts to shared
