+++ src/renderer/context/CurationContext.ts
@@ -4,7 +4,7 @@ import { CurationIndexImage, EditAddAppCurationMeta, EditCuration, EditCurationM
 import { createContextReducer } from '../context-reducer/contextReducer';
 import { ReducerAction } from '../context-reducer/interfaces';
 import { createCurationIndexImage } from '../curate/importCuration';
-import { uuid } from '../util/uuid';
+import { uuid } from '@shared/utils/uuid';
 
 const curationDefaultState: CurationsState = {
   defaultMetaData: undefined,
